requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.20
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# AI and ML dependencies
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.32.1
accelerate==0.22.0
safetensors==0.3.3
opencv-python==4.8.0.76
insightface==0.7.3
onnxruntime==1.15.1
tensorflow==2.13.0
keras==2.13.1

# Utilities
tqdm==4.66.1
numpy==1.24.4
scipy==1.11.2
matplotlib==3.7.2
pyyaml==6.0.1
